# Avoid compiling internal packages during installation.