--- a/graylog2-web-interface/packages/graylog-web-plugin/.gitignore
+++ b/graylog2-web-interface/packages/graylog-web-plugin/.gitignore
@@ -1,4 +1,3 @@
 node_modules
 npm-debug.log
-lib/**