--- a/graylog2-web-interface/packages/graylog-web-plugin/index.js
+++ b/graylog2-web-interface/packages/graylog-web-plugin/index.js
@@ -14,7 +14,8 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
+
 module.exports = {
-  PluginWebpackConfig: require('./lib/PluginWebpackConfig'),
-  loadBuildConfig: require('./lib/loadBuildConfig'),
+  PluginWebpackConfig: require('./src/PluginWebpackConfig'),
+  loadBuildConfig: require('./src/loadBuildConfig'),
 };