--- a/graylog2-web-interface/packages/graylog-web-plugin/package.json
+++ b/graylog2-web-interface/packages/graylog-web-plugin/package.json
@@ -3,14 +3,6 @@
   "version": "5.0.0-SNAPSHOT",
   "description": "Helper code for streamlining Graylog web interface plugin development",
   "main": "index.js",
-  "scripts": {
-    "test": "eslint src/*",
-    "build": "babel -d lib/ src/",
-    "preparebuild": "yarn install --ignore-scripts",
-    "compile": "yarn run preparebuild && yarn run build",
-    "prepublish": "yarn run compile",
-    "preinstall": "test -d src/ && yarn run compile"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Graylog2/graylog-web-plugin.git"
@@ -27,7 +19,7 @@
   },
   "homepage": "https://github.com/Graylog2/graylog-web-plugin#readme",
   "files": [
-    "lib/",
+    "src/",
     "templates/",
     "index.js",
     "plugin.js",
@@ -64,9 +56,5 @@
     "webpack-cleanup-plugin": "0.5.1",
     "webpack-cli": "4.10.0",
     "webpack-merge": "4.2.2"
-  },
-  "devDependencies": {
-    "@babel/cli": "7.16.8",
-    "babel-plugin-add-module-exports": "^1.0.4"
   }
 }