--- a/graylog2-web-interface/packages/graylog-web-plugin/src/PluginWebpackConfig.js
+++ b/graylog2-web-interface/packages/graylog-web-plugin/src/PluginWebpackConfig.js
@@ -1,4 +1,4 @@
- /*
+/*
  * Copyright (C) 2020 Graylog, Inc.
  *
  * This program is free software: you can redistribute it and/or modify