--- a/graylog2-web-interface/packages/jest-preset-graylog/.gitignore
+++ b/graylog2-web-interface/packages/jest-preset-graylog/.gitignore
@@ -1,3 +1,2 @@
 node_modules
 npm-debug.log
-lib/**