--- a/graylog2-web-interface/packages/jest-preset-graylog/lib/timeouts.js
+++ b/graylog2-web-interface/packages/jest-preset-graylog/lib/timeouts.js
@@ -16,6 +16,11 @@
  */
 const parsedTimeoutMultiplier = Number.parseFloat(process.env.TIMEOUT_MULTIPLIER);
 
-export const timeoutMultiplier = () => (Number.isFinite(parsedTimeoutMultiplier) ? parsedTimeoutMultiplier : 1.0);
+const timeoutMultiplier = () => (Number.isFinite(parsedTimeoutMultiplier) ? parsedTimeoutMultiplier : 1.0);
 
-export const applyTimeoutMultiplier = (x) => x * timeoutMultiplier();
+const applyTimeoutMultiplier = (x) => x * timeoutMultiplier();
+
+module.exports = {
+  timeoutMultiplier,
+  applyTimeoutMultiplier,
+};