--- a/graylog2-web-interface/packages/jest-preset-graylog/package.json
+++ b/graylog2-web-interface/packages/jest-preset-graylog/package.json
@@ -3,13 +3,6 @@
   "version": "1.0.0",
   "description": "Graylog jest preset, contains the jest configuration and dependencies to setup the test environment for the web interface",
   "main": "jest-preset.js",
-  "scripts": {
-    "build": "babel -d lib/ src/ --copy-files",
-    "preparebuild": "yarn install --ignore-scripts --production=false",
-    "compile": "yarn run preparebuild && yarn run build",
-    "prepublish": "yarn run compile",
-    "preinstall": "yarn run compile || echo \"src\" folder missing, skipping preinstall"
-  },
   "keywords": [
     "graylog",
     "jest"
@@ -34,12 +27,5 @@
     "jest-enzyme": "^7.1.2",
     "jest-styled-components": "7.1.1",
     "react-select-event": "^5.0.0"
-  },
-  "devDependencies": {
-    "@babel/cli": "7.16.8",
-    "@babel/core": "7.16.12",
-    "@babel/plugin-proposal-class-properties": "7.16.7",
-    "@babel/preset-env": "7.16.11",
-    "babel-plugin-add-module-exports": "^1.0.4"
   }
 }