## Description
<!--- Describe your changes in detail -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

In order to speed up package installation in our build, this PR is removing the most time-consuming step: transpiling our own internal packages (`graylog-web-plugin` / `jest-preset-graylog`).

For this, we do include these packages in our `babel` configs for transpilation during asset building/test runs (in `webpack`/`jest` configs).

This speeds up the process of `yarn install` significantly. Some numbers for a complete project build:

```
before:
graylog-project yarn install  230.40s user 110.82s system 79% cpu 7:08.55 total

after:
graylog-project yarn install  79.00s user 51.69s system 189% cpu 1:09.05 total
```

/nocl

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Refactoring (non-breaking change)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.